fixes lots of tools
