+++ main/app/Resources/modules/content/form/components/data.jsx
@@ -213,6 +213,7 @@ FormData.propTypes = {
 }
 
 FormData.defaultProps = {
+  level: 2,
   data: {}
 }
