+++ main/core/Resources/modules/tools/contact/components/visible-users.jsx
@@ -8,8 +8,8 @@ import {ListData} from '#/main/app/content/list/containers/data'
 import {constants as listConst} from '#/main/app/content/list/constants'
 
 import {trans} from '#/main/app/intl/translation'
-import {actions} from '#/main/core/user/contact/store'
-import {OptionsType} from '#/main/core/user/contact/prop-types'
+import {actions} from '#/main/core/tools/contact/store'
+import {OptionsType} from '#/main/core/tools/contact/prop-types'
 import {UserCard} from '#/main/core/user/data/components/user-card'
 
 const VisibleUsersComponent = props =>
