+++ main/core/Resources/modules/tools/contact/store/index.js
@@ -1,0 +1,11 @@
+/**
+ * Contact store.
+ */
+
+import {actions} from '#/main/core/tools/contact/store/actions'
+import {reducer} from '#/main/core/tools/contact/store/reducer'
+
+export {
+  actions,
+  reducer
+}
