+++ main/core/Resources/modules/tools/contact/store/reducer.js
