+++ main/core/Resources/modules/tools/home/components/tool.jsx
@@ -1,39 +1,34 @@
 import React from 'react'
 import {PropTypes as T} from 'prop-types'
-import {connect} from 'react-redux'
 
-import {Router, Routes, withRouter} from '#/main/app/router'
+import {Routes} from '#/main/app/router'
 
 import {Tab as TabTypes} from '#/main/core/tools/home/prop-types'
 import {Player} from '#/main/core/tools/home/player/components/player'
 import {Editor} from '#/main/core/tools/home/editor/components/editor'
-import {actions, selectors} from '#/main/core/tools/home/store'
-import {selectors as editorSelectors} from '#/main/core/tools/home/editor/store'
 
-const Tool = props =>
-  <Router>
-    <Routes
-      redirect={[
-        {from: '/', exact: true, to: '/tab/'+props.sortedTabs[0].id },
-        {from: '/edit', exact: true, to: '/edit/tab/'+props.editorTabs[0].id}
-      ]}
-      routes={[
-        {
-          path: '/tab/:id?',
-          exact: true,
-          component: Player,
-          onEnter: (params) => props.setCurrentTab(params.id)
-        }, {
-          path: '/edit/tab/:id?',
-          component: Editor,
-          onEnter: (params) => props.setCurrentTab(params.id),
-          disabled: !props.editable
-        }
-      ]}
-    />
-  </Router>
+const HomeTool = props =>
+  <Routes
+    redirect={[
+      {from: '/', exact: true, to: '/tab/'+props.sortedTabs[0].id },
+      {from: '/edit', exact: true, to: '/edit/tab/'+props.editorTabs[0].id}
+    ]}
+    routes={[
+      {
+        path: '/tab/:id?',
+        exact: true,
+        component: Player,
+        onEnter: (params) => props.setCurrentTab(params.id)
+      }, {
+        path: '/edit/tab/:id?',
+        component: Editor,
+        onEnter: (params) => props.setCurrentTab(params.id),
+        disabled: !props.editable
+      }
+    ]}
+  />
 
-Tool.propTypes = {
+HomeTool.propTypes = {
   sortedTabs: T.arrayOf(T.shape(
     TabTypes.propTypes
   )),
@@ -44,20 +39,6 @@ Tool.propTypes = {
   editable: T.bool.isRequired,
   setCurrentTab: T.func.isRequired
 }
-
-const HomeTool = withRouter(connect(
-  (state) => ({
-    editable: selectors.editable(state),
-    sortedTabs: selectors.sortedTabs(state),
-    editorTabs: editorSelectors.editorTabs(state),
-    currentTab: selectors.currentTab(state)
-  }),
-  (dispatch) => ({
-    setCurrentTab(tab){
-      dispatch(actions.setCurrentTab(tab))
-    }
-  })
-)(Tool))
 
 export {
   HomeTool
