--- main/core/Resources/modules/user/contact/store/index.js
@@ -1,11 +1,0 @@
-/**
- * Contact store.
- */
-
-import {actions} from '#/main/core/user/contact/store/actions'
-import {reducer} from '#/main/core/user/contact/store/reducer'
-
-export {
-  actions,
-  reducer
-}
