+++ main/core/Resources/modules/workspace/user/index.js
@@ -1,7 +1,7 @@
 import {bootstrap} from '#/main/app/dom/bootstrap'
 
 import {reducer} from '#/main/core/workspace/user/reducer'
-import {UserTool} from '#/main/core/workspace/user/components/tool'
+import {UserTool} from '#/main/core/workspace/user/containers/tool'
 
 // mount the react application
 bootstrap(
