+++ plugin/agenda/Resources/modules/tools/agenda/components/event.jsx
