+++ plugin/agenda/Resources/modules/tools/agenda/modals/event/index.js
@@ -6,7 +6,7 @@
 import {registry} from '#/main/app/modals/registry'
 
 // gets the modal component
-import {EventModal} from '#/plugin/agenda/components/modal/components/event'
+import {EventModal} from '#/plugin/agenda/tools/agenda/modals/event/components/modal'
 
 const MODAL_EVENT = 'MODAL_EVENT'
