+++ plugin/agenda/Resources/modules/tools/agenda/store/actions.js
