+++ plugin/agenda/Resources/modules/tools/agenda/store/index.js
@@ -1,0 +1,8 @@
+
+import {actions} from '#/plugin/agenda/tools/agenda/store/actions'
+import {reducer} from '#/plugin/agenda/tools/agenda/store/reducer'
+
+export {
+  actions,
+  reducer
+}
