+++ plugin/agenda/Resources/modules/tools/agenda/store/reducer.js
@@ -1,13 +1,15 @@
 import {makeReducer, combineReducers} from '#/main/app/store/reducer'
 import {makeFormReducer} from '#/main/app/content/form/store/reducer'
-import {AGENDA_UPDATE_FILTER_TYPE, AGENDA_UPDATE_FILTER_WORKSPACE} from '#/plugin/agenda/actions'
+
+import {
+  AGENDA_UPDATE_FILTER_TYPE,
+  AGENDA_UPDATE_FILTER_WORKSPACE
+} from '#/plugin/agenda/tools/agenda/store/actions'
 
 const reducer = {
-  current: makeFormReducer(
-    'events.current',
-    {},
-    {}
-  ),
+  current: makeFormReducer('events.current'),
+  workspace: makeReducer(null),
+  workspaces: makeReducer({}),
   filters: combineReducers({
     types: makeReducer(['event', 'task'], {
       [AGENDA_UPDATE_FILTER_TYPE] : (state, action) => action.filters
