+++ plugin/agenda/assets.json
@@ -1,7 +1,7 @@
 {
   "webpack": {
     "entry": {
-      "agenda": "index"
+      "agenda": "tools/agenda/index"
     }
   },
   "styles": {
