+++ plugin/drop-zone/Resources/modules/resources/dropzone/containers/resource.jsx
@@ -1,5 +1,6 @@
 import {connect} from 'react-redux'
 
+import {withRouter} from '#/main/app/router'
 import {withReducer} from '#/main/app/store/components/withReducer'
 
 import {selectors as resourceSelect} from '#/main/core/resource/store'
@@ -11,23 +12,25 @@ import {reducer, select} from '#/plugin/drop-zone/resources/dropzone/store'
 import {actions as playerActions} from '#/plugin/drop-zone/resources/dropzone/player/actions'
 import {actions as correctionActions} from '#/plugin/drop-zone/resources/dropzone/correction/actions'
 
-const DropzoneResource = withReducer(select.STORE_NAME, reducer)(
-  connect(
-    (state) => ({
-      canEdit: hasPermission('edit', resourceSelect.resourceNode(state)),
-      dropzone: select.dropzone(state),
-      myDrop: select.myDrop(state)
-    }),
-    (dispatch) => ({
-      resetForm: (formData) => dispatch(formActions.resetForm(select.STORE_NAME+'.dropzoneForm', formData)),
+const DropzoneResource = withRouter(
+  withReducer(select.STORE_NAME, reducer)(
+    connect(
+      (state) => ({
+        canEdit: hasPermission('edit', resourceSelect.resourceNode(state)),
+        dropzone: select.dropzone(state),
+        myDrop: select.myDrop(state)
+      }),
+      (dispatch) => ({
+        resetForm: (formData) => dispatch(formActions.resetForm(select.STORE_NAME+'.dropzoneForm', formData)),
 
-      fetchDrop: (dropId, type) => dispatch(correctionActions.fetchDrop(dropId, type)),
-      resetCurrentDrop: () => dispatch(correctionActions.resetCurrentDrop()),
-      fetchCorrections: (dropzoneId) => dispatch(correctionActions.fetchCorrections(dropzoneId)),
-      resetCorrectorDrop: () => dispatch(correctionActions.resetCorrectorDrop()),
-      fetchPeerDrop: () => dispatch(playerActions.fetchPeerDrop())
-    })
-  )(DropzoneResourceComponent)
+        fetchDrop: (dropId, type) => dispatch(correctionActions.fetchDrop(dropId, type)),
+        resetCurrentDrop: () => dispatch(correctionActions.resetCurrentDrop()),
+        fetchCorrections: (dropzoneId) => dispatch(correctionActions.fetchCorrections(dropzoneId)),
+        resetCorrectorDrop: () => dispatch(correctionActions.resetCorrectorDrop()),
+        fetchPeerDrop: () => dispatch(playerActions.fetchPeerDrop())
+      })
+    )(DropzoneResourceComponent)
+  )
 )
 
 export {
