+++ plugin/team/Resources/modules/tools/team/components/tool.jsx
@@ -1,12 +1,7 @@
 import React from 'react'
 import {PropTypes as T} from 'prop-types'
-import {connect} from 'react-redux'
 
-import {actions as formActions} from '#/main/app/content/form/store'
-
-import {makeId} from '#/main/core/scaffolding/id'
 import {trans} from '#/main/app/intl/translation'
-import {select as workspaceSelect} from '#/main/core/workspace/selectors'
 import {
   PageContainer,
   PageHeader,
@@ -16,7 +11,6 @@ import {
 } from '#/main/core/layout/page'
 import {RoutedPageContent} from '#/main/core/layout/router/components/page'
 
-import {selectors, actions} from '#/plugin/team/tools/team/store'
 import {TeamParams as TeamParamsType} from '#/plugin/team/tools/team/prop-types'
 import {Editor} from '#/plugin/team/tools/team/components/editor'
 import {Teams} from '#/plugin/team/tools/team/components/teams'
@@ -24,7 +18,7 @@ import {Team} from '#/plugin/team/tools/team/components/team'
 import {TeamForm} from '#/plugin/team/tools/team/components/team-form'
 import {MultipleTeamForm} from '#/plugin/team/tools/team/components/multiple-team-form'
 
-const TeamToolComponent = props =>
+const TeamTool = props =>
   <PageContainer>
     <PageHeader title={trans('team', {}, 'team')}>
       {props.canEdit ?
@@ -115,7 +109,7 @@ const TeamToolComponent = props =>
     />
   </PageContainer>
 
-TeamToolComponent.propTypes = {
+TeamTool.propTypes = {
   canEdit: T.bool.isRequired,
   teamParams: T.shape(TeamParamsType.propTypes).isRequired,
   resourceTypes: T.arrayOf(T.string).isRequired,
@@ -126,51 +120,6 @@ TeamToolComponent.propTypes = {
   openMultipleTeamsForm: T.func.isRequired,
   resetMultipleTeamsForm: T.func.isRequired
 }
-
-const TeamTool = connect(
-  (state) => ({
-    canEdit: selectors.canEdit(state),
-    teamParams: selectors.teamParams(state),
-    resourceTypes: selectors.resourceTypes(state),
-    workspaceId: workspaceSelect.workspace(state).uuid
-  }),
-  (dispatch) => ({
-    resetForm(formData) {
-      dispatch(formActions.resetForm('teamParamsForm', formData))
-    },
-    openCurrentTeam(id, teamParams, workspaceId, resourceTypes) {
-      const defaultValue = {
-        id: makeId(),
-        workspace: {
-          uuid: workspaceId
-        },
-        selfRegistration: teamParams.selfRegistration,
-        selfUnregistration: teamParams.selfUnregistration,
-        publicDirectory: teamParams.publicDirectory,
-        deletableDirectory: teamParams.deletableDirectory,
-        creatableResources: resourceTypes
-      }
-      dispatch(actions.openForm('teams.current', id, defaultValue))
-    },
-    resetCurrentTeam() {
-      dispatch(formActions.resetForm('teams.current', {}, true))
-    },
-    openMultipleTeamsForm(teamParams, resourceTypes) {
-      const defaultValue = {
-        nbTeams: 1,
-        selfRegistration: teamParams.selfRegistration,
-        selfUnregistration: teamParams.selfUnregistration,
-        publicDirectory: teamParams.publicDirectory,
-        deletableDirectory: teamParams.deletableDirectory,
-        creatableResources: resourceTypes
-      }
-      dispatch(formActions.resetForm('teams.multiple', defaultValue, true))
-    },
-    resetMultipleTeamsForm() {
-      dispatch(formActions.resetForm('teams.multiple', {}, true))
-    }
-  })
-)(TeamToolComponent)
 
 export {
   TeamTool
