+++ plugin/team/Resources/modules/tools/team/index.js
@@ -1,8 +1,7 @@
 import {reducer} from '#/plugin/team/tools/team/store'
-import {TeamTool} from '#/plugin/team/tools/team/components/tool'
+import {TeamTool} from '#/plugin/team/tools/team/containers/tool'
 
 export const App = () => ({
   component: TeamTool,
-  store: reducer,
-  initialData: (initialData) => Object.assign({}, initialData)
+  store: reducer
 })
